backend/pyproject.toml
[tool.poetry]
name = "graphrag-exam-evaluation-api"
version = "1.0.0"
description = "AI-Powered Exam Correction System for Primary School Math using Neo4j GraphRAG"
authors = ["Your Team <team@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.109.0"
uvicorn = {extras = ["standard"], version = "^0.27.0"}
neo4j = "^5.16.0"
openai = "^1.12.0"
python-multipart = "^0.0.6"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
pydantic = {extras = ["email"], version = "^2.5.0"}
pydantic-settings = "^2.1.0"
python-dotenv = "^1.0.0"
pypdf2 = "^3.0.1"
pdf2image = "^1.17.0"
pillow = "^10.2.0"
pytesseract = "^0.3.10"
langchain = ">=0.1.0"
langchain-openai = ">=0.0.5"
langchain-community = ">=0.0.19"
tiktoken = "^0.5.2"
prometheus-client = "^0.19.0"
structlog = "^24.1.0"
httpx = "^0.26.0"
tenacity = "^8.2.3"
aiofiles = "^23.2.1"
numpy = "^1.26.0"
sympy = "^1.12"  # For mathematical expression evaluation

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.4"
pytest-asyncio = "^0.23.3"
pytest-cov = "^4.1.0"
black = "^24.1.1"
ruff = "^0.1.14"
mypy = "^1.8.0"
pre-commit = "^3.6.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py311']

[tool.ruff]
line-length = 100
select = ["E", "F", "I"]
ignore = ["E501"]
